# perf(storage): simplify table watermark index