--- a/src/storage/Cargo.toml
+++ b/src/storage/Cargo.toml
@@ -145,5 +145,9 @@ harness = false
 name = "bench_imm_compact"
 harness = false
 
+[[bench]]
+name = "bench_table_watermarks"
+harness = false
+
 [lints]
 workspace = true