--- a/src/storage/hummock_test/src/hummock_storage_tests.rs
+++ b/src/storage/hummock_test/src/hummock_storage_tests.rs
@@ -1,4 +1,3 @@
-use std::collections::VecDeque;
 // Copyright 2024 RisingWave Labs
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -29,7 +28,7 @@ use risingwave_hummock_sdk::key::{
     gen_key_from_bytes, prefixed_range_with_vnode, FullKey, TableKey, UserKey, TABLE_PREFIX_LEN,
 };
 use risingwave_hummock_sdk::table_watermark::{
-    visible_watermark, VnodeWatermark, WatermarkDirection,
+    TableWatermarksIndex, VnodeWatermark, WatermarkDirection,
 };
 use risingwave_hummock_sdk::EpochWithGap;
 use risingwave_rpc_client::HummockMetaClient;
@@ -2187,19 +2186,23 @@ async fn test_table_watermark() {
     let (local1, local2) = test_after_epoch2(local1, local2).await;
 
     let check_version_table_watermark = |version: PinnedVersion| {
-        let table_watermarks = version
-            .version()
-            .table_watermarks
-            .get(&TEST_TABLE_ID)
-            .unwrap();
-        assert_eq!(WatermarkDirection::Ascending, table_watermarks.direction);
+        let table_watermarks = TableWatermarksIndex::new_committed(
+            version
+                .version()
+                .table_watermarks
+                .get(&TEST_TABLE_ID)
+                .unwrap()
+                .clone(),
+            version.max_committed_epoch(),
+        );
+        assert_eq!(WatermarkDirection::Ascending, table_watermarks.direction());
         assert_eq!(
             gen_inner_key(watermark1),
-            visible_watermark(vnode1, epoch1, &VecDeque::new(), Some(table_watermarks)).unwrap()
+            table_watermarks.read_watermark(vnode1, epoch1).unwrap()
         );
         assert_eq!(
             gen_inner_key(watermark1),
-            visible_watermark(vnode2, epoch1, &VecDeque::new(), Some(table_watermarks)).unwrap()
+            table_watermarks.read_watermark(vnode2, epoch1).unwrap()
         );
     };