--- a/src/storage/src/hummock/local_version/pinned_version.rs
+++ b/src/storage/src/hummock/local_version/pinned_version.rs
@@ -19,7 +19,6 @@ use std::time::{Duration, Instant};
 
 use auto_enums::auto_enum;
 use risingwave_common::catalog::TableId;
-use risingwave_hummock_sdk::table_watermark::TableWatermarksIndex;
 use risingwave_hummock_sdk::version::HummockVersion;
 use risingwave_hummock_sdk::{CompactionGroupId, HummockVersionId, INVALID_VERSION_ID};
 use risingwave_pb::hummock::hummock_version::Levels;
@@ -77,7 +76,6 @@ impl Drop for PinnedVersionGuard {
 pub struct PinnedVersion {
     version: Arc<HummockVersion>,
     compaction_group_index: Arc<HashMap<TableId, CompactionGroupId>>,
-    table_watermark_index: Arc<HashMap<TableId, TableWatermarksIndex>>,
     guard: Arc<PinnedVersionGuard>,
 }
 
@@ -88,12 +86,10 @@ impl PinnedVersion {
     ) -> Self {
         let version_id = version.id;
         let compaction_group_index = version.build_compaction_group_info();
-        let table_watermark_index = version.build_table_watermarks_index();
 
         PinnedVersion {
             version: Arc::new(version),
             compaction_group_index: Arc::new(compaction_group_index),
-            table_watermark_index: Arc::new(table_watermark_index),
             guard: Arc::new(PinnedVersionGuard::new(
                 version_id,
                 pinned_version_manager_tx,
@@ -105,10 +101,6 @@ impl PinnedVersion {
         self.compaction_group_index.clone()
     }
 
-    pub fn table_watermark_index(&self) -> &Arc<HashMap<TableId, TableWatermarksIndex>> {
-        &self.table_watermark_index
-    }
-
     pub(crate) fn new_pin_version(&self, version: HummockVersion) -> Self {
         assert!(
             version.id >= self.version.id,
@@ -118,12 +110,10 @@ impl PinnedVersion {
         );
         let version_id = version.id;
         let compaction_group_index = version.build_compaction_group_info();
-        let table_watermark_index = version.build_table_watermarks_index();
 
         PinnedVersion {
             version: Arc::new(version),
             compaction_group_index: Arc::new(compaction_group_index),
-            table_watermark_index: Arc::new(table_watermark_index),
             guard: Arc::new(PinnedVersionGuard::new(
                 version_id,
                 self.guard.pinned_version_manager_tx.clone(),