--- a/src/storage/src/hummock/store/hummock_storage.rs
+++ b/src/storage/src/hummock/store/hummock_storage.rs
@@ -12,6 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::collections::VecDeque;
 use std::future::Future;
 use std::ops::{Bound, Deref};
 use std::sync::atomic::{AtomicU64, Ordering as MemOrdering};
@@ -28,6 +29,7 @@ use risingwave_common_service::observer_manager::{NotificationClient, ObserverMa
 use risingwave_hummock_sdk::key::{
     is_empty_key_range, vnode, vnode_range, TableKey, TableKeyRange,
 };
+use risingwave_hummock_sdk::table_watermark::rewrite_range_with_table_watermark;
 use risingwave_hummock_sdk::HummockReadEpoch;
 use risingwave_pb::hummock::SstableInfo;
 use risingwave_rpc_client::HummockMetaClient;
@@ -123,8 +125,14 @@ pub fn get_committed_read_version_tuple(
     mut key_range: TableKeyRange,
     epoch: HummockEpoch,
 ) -> (TableKeyRange, ReadVersionTuple) {
-    if let Some(index) = version.table_watermark_index().get(&table_id) {
-        index.rewrite_range_with_table_watermark(epoch, &mut key_range)
+    if let Some(table_watermarks) = version.version().table_watermarks.get(&table_id) {
+        rewrite_range_with_table_watermark(
+            epoch,
+            table_watermarks.direction,
+            &mut key_range,
+            &VecDeque::new(),
+            Some(table_watermarks),
+        )
     }
     (key_range, (vec![], vec![], version))
 }