@wenym1 2024-03-28
Added benchmark.

Benchmark result is as followed. The test setting is, 500 epochs between safe epoch and committed epoch, and 500 staging epochs. Vnodes are partitioned into 16 parts.

This branch
```
new pinned version      time:   [37.867 µs 38.279 µs 38.689 µs]
                        change: [-100.000% -100.000% -100.000%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  3 (3.00%) high mild
  4 (4.00%) high severe

apply committed watermark
                        time:   [161.95 µs 167.16 µs 172.45 µs]
                        change: [-99.989% -99.989% -99.988%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild

read latest watermark   time:   [2.4513 µs 2.5030 µs 2.6131 µs]
                        change: [+2.5368% +3.8159% +6.2119%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe

read committed watermark
                        time:   [81.068 µs 81.169 µs 81.271 µs]
                        change: [-0.2821% -0.0573% +0.1584%] (p = 0.61 > 0.05)
                        No change in performance detected.

```

main branch
```
read latest watermark   time:   [2.3970 µs 2.4011 µs 2.4056 µs]
                        change: [-85.105% -85.063% -85.021%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild

read committed watermark
                        time:   [81.021 µs 81.195 µs 81.395 µs]
                        change: [+706.95% +709.75% +712.35%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) high mild
  2 (2.00%) high severe
```

The time for `build_index` and `apply_committed_watermarks` can be ignored in this branch compared to the main branch, because there is no index to maintain. The time in the main branch is so great that we had to reduce the sample size to be able to finish the test in reasonable time.

For query, the time to query the latest epoch is 85% lower than the main branch, and the time to query a middle epoch is 7 times the main branch. This is because the implementation is linearly searching from latest epoch to old epoch, and therefore is't biased for querying the watermark in the latest epoch.

In conclusion, compared to the main branch, this PR save the significant time to build and maintain the table watermark index. In terms of query, this branch is even faster than the main branch in querying the latest epoch. It only performs poorly in querying a mvcc epoch, which is not a common use case.